Write tests for passing options and code coverage
The tests are a little contrived, but should show an example of using options for passing select options
to the child fields (with defaults). It also shows how to use the form factory to pass options to a form
and configure a child field with it.

At time of commit, all tests pass on PHP 5.6.33 and PHPUnit 5.7.26.